Created FlagCollection and did general pimping
Created "sidebar.html" to replace "index.html" as the sidebar where all
scripts are being listed. Now index can be re-purposed as a front page
for the project/library.

Added two classes; "FlagCollection" and "FlagSource". The first one is a
collection of flags, one is kept in every instance of every class that
uses flags, it stores them and fetches them. The second should there
only be one instance of per type of flag collection (scripts, help,
etc.) and it will keep the default flag values and the def-set values.

Added "*.sass-cache" to .gitignore. The sass compiler I use will create
a folder just that if it doesn't have admin priveliges.

Minor changes to the documentation templates.